microsservices/requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.1.0
celery==4.4.0
Django==3.0.2
django-celery-results==1.2.0
django-model-utils==3.2.0
djangorestframework==3.11.0
importlib-metadata==1.5.0
kombu==4.6.7
psycopg2-binary==2.8.4
python-decouple==3.3
pytz==2019.3
redis==3.2.0
sqlparse==0.3.0
vine==1.3.0
zipp==2.1.0
